Fix paged queue size issue (merge main -> ce/main 111015)
[git-p4: depot-paths = "//dev/coherence-ce/main/": change = 111041]

+++ prj/coherence-core/src/main/java/com/tangosol/internal/net/queue/paged/PagedQueueKey.java
@@ -37,7 +37,7 @@ public class PagedQueueKey
      * @param bucketId  the id of the bucket this key belongs to
      * @param elementId the id of the element within the bucket
      */
-    public PagedQueueKey(int bucketId, int elementId)
+    public PagedQueueKey(int bucketId, long elementId)
         {
         super(bucketId, elementId);
         }
@@ -86,15 +86,6 @@ public class PagedQueueKey
         }
 
     // ----- constants ------------------------------------------------------
-//
-//    /** The POF id of the bucket id */
-//    public static final int POF_BUCKET_ID = 0;
-//
-//    /** The POF id of the element id */
-//    public static final int POF_ELEMENT_ID = 1;
-//
-//    /** static helper field for navigating to the bucket id from a PofValue */
-//    public static final PofNavigator BUCKET_ID_NAVIGATOR  = new SimplePofPath(POF_BUCKET_ID);
 
     /** static helper field for extracting the bucket id from a cache entry */
     public static final UniversalExtractor<PagedQueueKey, Integer> BUCKET_ID_EXTRACTOR
